requirements.txt
pandas
numpy
sqlalchemy
faiss-cpu
sentence-transformers
openai
gradio
tqdm
matplotlib
plotly
python-dotenv
rapidfuzz
